Move formatAttributesForDisplay function to separate utils.ts file
