+++ diag/app/package.json
@@ -51,6 +51,6 @@
   "homepage": ".",
   "devDependencies": {
     "@types/react-router-bootstrap": "^0.24.5",
-    "rimraf": "3.0.2"
+    "rimraf": "^3.0.2"
   }
 }
